Conflicts:
	src/main/java/org/bukkit/Color.java
	src/main/java/org/bukkit/configuration/serialization/ConfigurationSerialization.java

+++ src/main/java/org/bukkit/Color.java
@@ -279,10 +279,10 @@ public final class Color implements ConfigurationSerializable {
             totalMax += Math.max(Math.max(color.getRed(), color.getGreen()), color.getBlue());
         }
 
-        int averageRed = totalRed / (colors.length + 1);
-        int averageGreen = totalGreen / (colors.length + 1);
-        int averageBlue = totalBlue / (colors.length + 1);
-        int averageMax = totalMax / (colors.length + 1);
+        float averageRed = (float) totalRed / (colors.length + 1);
+        float averageGreen = (float) totalGreen / (colors.length + 1);
+        float averageBlue = (float) totalBlue / (colors.length + 1);
+        float averageMax = (float) totalMax / (colors.length + 1);
 
         float maximumOfAverages = Math.max(Math.max(averageRed, averageGreen), averageBlue);
         float gainFactor = averageMax / maximumOfAverages;
